added switch specifics
