+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b1c0cfbb2de2842fb90ebe2912fcdc35
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
