+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch/SwitchCacheNone.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SwitchCacheNone : Scenario
+{
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        Configuration.SwitchCacheType = BugsnagUnity.SwitchCacheType.None;
+    }
+
+    public override void Run()
+    {
+        DoSimpleNotify("SwitchCacheNone");
+    }
+}
