+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch/SwitchCacheNone.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e017dbbe548946338fe06f8e85d6781
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
