+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch/SwitchMaxCacheSize.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 02e93ca43bfe04470be8fd9e0e49db13
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
