+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch/SwitchMetadata.cs
@@ -1,0 +1,7 @@
+﻿public class SwitchMetadata : Scenario
+{
+    public override void Run()
+    {
+        DoSimpleNotify("SwitchMetadata");
+    }
+}
