+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch/SwitchMetadata.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1f2a957475664c938a1b5abc2d9a67c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
