+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch/SwitchPersistEvent.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SwitchPersistEvent : Scenario
+{
+    public override void PrepareConfig(string apiKey, string host)
+    {
+        base.PrepareConfig(apiKey, host);
+        SetInvalidEndpoints();
+    }
+
+
+}
