+++ features/fixtures/maze_runner/Assets/Scripts/Scenarios/Switch/SwitchPersistEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0b3e3ed23947246a6a870a949fd38633
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
