+++ features/switch/switch.feature
@@ -1,8 +1,8 @@
 Feature: Switch Specific Tests
 
   Scenario: SwitchCacheType set to None
-    When I run the game in the "PersistEvent" state
-    And I wait for 5 seconds
+    When I run the game in the "SwitchPersistEvent" state
+    And I wait for requests to fail
     And I close the Unity app
     And I run the game in the "SwitchCacheNone" state
     And I wait to receive an error
@@ -13,17 +13,17 @@ Feature: Switch Specific Tests
 
   Scenario: Max Cache Size
     When I run the game in the "MaxSwitchCacheSize" state
-    And I wait for 10 seconds
+    And I wait for requests to fail
     And I close the Unity app
-    And I run the game in the "(noop)" state
+    And I run the game in the "StartSDKDefault" state
     And I wait to receive 1 errors
     And the exception "message" equals "LARGE PAYLOAD 2"
 
   Scenario: Switch Metadata
-    When I run the game in the "UncaughtException" state
+    When I run the game in the "SwitchMetadata" state
     And I wait to receive an error
     Then the error is valid for the error reporting API sent by the Unity notifier
-    And the exception "errorClass" equals "ExecutionEngineException"
+    And the exception "message" equals "SwitchMetadata"
     And the event "app.type" equals "nintendo-switch"
     And the event "device.osName" equals "Nintendo Switch"
     And the event "device.model" equals "Switch"
